@@ -4,13 +4,13 @@ version = "0.1.0"
 description = "Generative Agents: Interactive Simulacra of Human Behavior"
 readme = "README.md"
 license = { text = "Apache-2.0" }
-requires-python = ">=3.9,<3.13"
+requires-python = ">=3.12"
 
 dependencies = [
-    "Django>=2.2,<3.0",
-    "django-cors-headers>=2.5,<4.0",
-    "django-storages-redux>=1.3,<2.0",
-    "numpy>=1.19.5",
+    "Django>=4.2.8,<5.0",
+    "django-cors-headers>=4.0,<5.0",
+    "django-storages>=1.14,<2.0",
+    "numpy>=1.26,<2.0",
     "openai>=0.27.0,<1.0",
     "python-dotenv>=1.2.1",
     "selenium>=4.8,<5.0",
@@ -18,15 +18,15 @@ dependencies = [
 
 [project.optional-dependencies]
 analysis = [
-    "gensim>=3.8,<4.0",
+    "gensim>=4.3.3,<5.0",
     "nltk>=3.6,<4.0",
-    "scipy>=1.5,<1.12",
-    "scikit-learn>=0.24,<2.0",
-    "matplotlib>=3.3,<4.0",
-    "pandas>=1.1,<3.0",
+    "scipy>=1.11,<2.0",
+    "scikit-learn>=1.3,<2.0",
+    "matplotlib>=3.8,<4.0",
+    "pandas>=2.0,<3.0",
     "seaborn>=0.12,<1.0",
-    "statsmodels>=0.13,<1.0",
-    "yellowbrick>=1.3,<2.0",
+    "statsmodels>=0.14,<1.0",
+    "yellowbrick>=1.5,<2.0",
     "scikit-posthocs>=0.7,<1.0",
     "trueskill>=0.4,<1.0",
 ]
