@@ -1,7 +1,7 @@
 """frontend_server URL Configuration
 
 The `urlpatterns` list routes URLs to views. For more information please see:
-    https://docs.djangoproject.com/en/2.2/topics/http/urls/
+    https://docs.djangoproject.com/en/4.2/topics/http/urls/
 Examples:
 Function views
     1. Add an import:  from my_app import views
@@ -13,8 +13,7 @@ Including another URLconf
     1. Import the include() function: from django.urls import include, path
     2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
-from django.conf.urls import include, url
-from django.urls import path
+from django.urls import include, path, re_path
 from django.contrib import admin
 from django.conf import settings
 from django.conf.urls.static import static
@@ -22,14 +21,14 @@ from django.conf.urls.static import static
 from translator import views as translator_views
 
 urlpatterns = [
-    url(r'^$', translator_views.landing, name='landing'),
-    url(r'^simulator_home$', translator_views.home, name='home'),
-    url(r'^demo/(?P<sim_code>[\w-]+)/(?P<step>[\w-]+)/(?P<play_speed>[\w-]+)/$', translator_views.demo, name='demo'),
-    url(r'^replay/(?P<sim_code>[\w-]+)/(?P<step>[\w-]+)/$', translator_views.replay, name='replay'),
-    url(r'^replay_persona_state/(?P<sim_code>[\w-]+)/(?P<step>[\w-]+)/(?P<persona_name>[\w-]+)/$', translator_views.replay_persona_state, name='replay_persona_state'),
-    url(r'^process_environment/$', translator_views.process_environment, name='process_environment'),
-    url(r'^update_environment/$', translator_views.update_environment, name='update_environment'),
-    url(r'^path_tester/$', translator_views.path_tester, name='path_tester'),
-    url(r'^path_tester_update/$', translator_views.path_tester_update, name='path_tester_update'),
+    re_path(r'^$', translator_views.landing, name='landing'),
+    re_path(r'^simulator_home$', translator_views.home, name='home'),
+    re_path(r'^demo/(?P<sim_code>[\w-]+)/(?P<step>[\w-]+)/(?P<play_speed>[\w-]+)/$', translator_views.demo, name='demo'),
+    re_path(r'^replay/(?P<sim_code>[\w-]+)/(?P<step>[\w-]+)/$', translator_views.replay, name='replay'),
+    re_path(r'^replay_persona_state/(?P<sim_code>[\w-]+)/(?P<step>[\w-]+)/(?P<persona_name>[\w-]+)/$', translator_views.replay_persona_state, name='replay_persona_state'),
+    re_path(r'^process_environment/$', translator_views.process_environment, name='process_environment'),
+    re_path(r'^update_environment/$', translator_views.update_environment, name='update_environment'),
+    re_path(r'^path_tester/$', translator_views.path_tester, name='path_tester'),
+    re_path(r'^path_tester_update/$', translator_views.path_tester_update, name='path_tester_update'),
     path('admin/', admin.site.urls),
 ]
